Fix burn (3)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@92999 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/DevVirtioNet_1_0.cpp
@@ -1890,7 +1890,7 @@ static int virtioNetR3RxPktMultibufXfer(PPDMDEVINS pDevIns, PVIRTIONET pThis, ui
     Log7Func(("  Sending packet header to guest...\n"));
 
     /* Copy packet header to rx buf provided by caller. */
-    uint32_t cbHdrEnqueued = pVirtqBuf->cbPhysReturn == cbPktHdr ? cbPktHdr : 0;
+    size_t cbHdrEnqueued = pVirtqBuf->cbPhysReturn == cbPktHdr ? cbPktHdr : 0;
     virtioCoreR3VirtqUsedBufPut(pDevIns, &pThis->Virtio, pRxVirtq->uIdx, cbPktHdr, pRxPktHdr, pVirtqBuf, cbHdrEnqueued);
 
     /* Cache address of uNumBuffers field of pkthdr to update ex post facto */
